refactor: promote errordef from manager to common (#67)
Move internal/manager/errordef to internal/common/errordef so common
packages can share coded errors without importing the manager layer, and
add ImagePullFailed / ContainerCreateFailed codes for the workload contract.
