+++ internal/manager/repository/db/run.go
@@ -12,7 +12,7 @@ import (
 	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/preset"
 	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/spec"
 	"github.com/seedspirit/nano-backend.ai/internal/common/encoding"
-	"github.com/seedspirit/nano-backend.ai/internal/manager/errordef"
+	"github.com/seedspirit/nano-backend.ai/internal/common/errordef"
 	"github.com/seedspirit/nano-backend.ai/internal/manager/repository/db/entity"
 )
